Inline Rust code again
After trying out a lot of things, I realized that the current version of mdBook seems to have a bug in `mdbook test` where it can't find the correct files (it works with master branch).
In addition I understand now that mdbook test is not executing Rust tests, but the Rust code itself, basically making sure that your written code compiles. Which explains why it did not pick up my tests.

As such, I decided to simplify the book structure again.